requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bottle==0.13.2
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cssselect2==0.7.0
fonttools==4.55.6
html5lib==1.1
idna==3.10
lxml==5.3.0
oscrypto==1.3.0
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.2.0
pyphen==0.17.2
python-bidi==0.6.3
python-dotenv==0.21.0
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
-e git+https://github.com/xhtml2pdf@d97ed6a99af0e4cd6c1e79d2cbe468db7f6d7138#egg=xhtml2pdf
zopfli==0.2.3.post1
